Update icon colour to inherit
The IE8 icon fallback uses a HTML Arrow Entity as various CSS properties that create the icon are not supported (eg transform) , this is added via the content property. This allows an older browser like IE8 to render an icon (and while, visually, less than ideal) that provides more affordance to users about the interaction.
 
While border-colour is more explicit to in order to style the icon, for this element to inherit the colouring correctly without additional code specially to target IE8 - colour is a more suitable property.  This allows both to IE8 and non-IE8 icons to receive the correct colour(s) while being interacted across various states (hover, focus)

The icon approach - is arguably a little more inline with graceful degradation principles