Update icons to relative values
- Using "govuk-em"
- Remove explicitly defined colour as inherited.  

Many users alter their browser settings -  a user might zoom in or increase the size of the font to adjust a webpage for their needs.

In order to more closely meet Accessibility Acceptable Criteria of "zoom-text-only 300%"[1] the icon within the Accordion has been adjusted to use relative[2] values utilising an existing mixin[5]

This allows the icon to scale correctly as zoom via "text only" is being used.  As it's a relative value (not a pixel) this is more inline with Responsive Design[3] principles and also more closely aligns with WCAG 1.4.4[4] and previous iteration[6]
---

[1] https://support.mozilla.org/en-US/kb/font-size-and-zoom-increase-size-of-web-pages#w_how-to-only-change-the-size-of-the-text

[2] https://thecssworkshop.com/lessons/relative-units

[3] https://alistapart.com/article/responsive-web-design/

[4] https://www.w3.org/TR/UNDERSTANDING-WCAG20/visual-audio-contrast-scale.html

[5] https://github.com/alphagov/govuk-frontend/blob/2f96f0d2cf088058d2beac7112167605c815b4df/src/govuk/tools/_px-to-em.scss#L12

[6] https://github.com/alphagov/govuk_publishing_components/pull/1884

---
Non exhausted examples of how to test this:

Within FireFox: 
View > Zoom > Zoom text only
Within Chrome
https://support.google.com/chrome/answer/96810?hl=en