Update to new Design
- Update Design

Addressing WCAG fails: 1.3.1 (Info and relationships)[1] 
The current Headings look like links but are Headings, they are also treated like buttons. This visual update more closely aligns with what the element is semantically with it's visual representation - in turn resolves this WCAG issue.

The original "+" icon being on the right-hand side could easily be missed by users that are zoomed-in - Also the type of icon is improved to indicate the type of interaction (open / close) that is expected upon interaction.

- Add Show all button

- Add Hidden "this section" > Show[ this section]
(@nnagewad) "We believed that there is enough visual affordance for users using the Chevron icon and just "Show" text[6]."

In order to improve this experience for AT users ("Show" what?) a hidden "this section" was added. There was an question (that requires user testing) whether users that see "Show" *and* hear "this section" would be impacted negativity.

- Add Wrapper for all inner "[icon] Show [this section]" elements called toggle text

- Instead of adding "govuk-heading-m" to the template file the relevant properties have been ported into the Accordion[7]

- Add focus states around inner Content
In order for the focus states to match the new Design and also be fluid so it responses to the inner Content an additional wrapper of "govuk-accordion__section-heading-focus-wrapper" has been added

----
Design update comments: 

(@christopherthomasdesign) "Padding was added to be similar to that of a button, a larger touch area and focus indicator compared to treating the element like a link alone.  We also explored making it look more like a secondary button but felt it look too separately from the rest of the component"
 
"Font size went from 16px to 19px of the Show / Show All Buttons, changed this in response to feedback about the 'show all' touch area not being big enough – not a perfect solution but does increase the clickable/touch area a bit"

"we’re increasing the padding to make the bottom of each section v distinct from the next section"

- Wider Context

Also addresses a WCAG 3.2.4[2] fail - there are multiple versions of an Accordion pattern in use (on GOV.UK Publishing not the GOV.UK Design System / Frontend) 

Eg "Step by Step")[3] uses an Accordion pattern - this inconsistency is ultimately is confusing for users and a WCAG fail. This Design iteration addresses a wider concern to provide a consistent Design language for such instances of this "Open / Close" pattern.

There are other instances of an Accordion Design that might not be explicitly clear are a type of Accordion.  For example a mobile-menu navigation pattern that reveals more items is a type of Accordion[4]. A filter menu is a type of Accordion[5])
----

[1]https://www.w3.org/TR/UNDERSTANDING-WCAG20/content-structure-separation-programmatic.html

[2] https://www.w3.org/WAI/WCAG21/Understanding/consistent-identification.html

[3] https://components.publishing.service.gov.uk/component-guide/step_by_step_nav

[4] https://github.com/alphagov/govuk_publishing_components/blob/873b80980be9a0f36cd87ba1b04ed218f78a75f2/app/views/govuk_publishing_components/components/_government_navigation.html.erb

[5] https://www.gov.uk/search/all?keywords=test&order=relevance

[6] https://lawsofux.com/aesthetic-usability-effect/

[7] https://github.com/alphagov/govuk-frontend/blob/7557564ecc8dbf6fbaf6b63c28dbc29e57239b77/src/govuk/core/_typography.scss#L391