Prevent panel text from clipping on mobile zoom
This clipping is more likely to happen when user increases text size on a mobile eg. using
iOS Safari text resize controls, and is more applicable to the panel than other components
since the padding of the panel component limits the width of text more than our other
components do.

Decrease panel padding and heading font size on mobile to allow the text flow better. Also
decrease content font size to reflect font size drop for headings.

As an if-all-else-fails attempt to stop very long words from clipping, force them to break
and wrap instead. Hopefully this is not something that would happen often since we now also
reduce the padding and font size to help with this.